Update Solver.java
Read your source code on "Kakurasu.java", and about how the game is supposed to be played.  Realized for each box, specified in a cols and rows coordinate, we first need to calculate how many combinations can be driven.

Made an ArrayList to calculate if the value is within 1-9.  
If there are two empty boxes, one number can be 1, and the other is (x-1)  such that x = a number between 1 through 9.

Hope this helps with your solver function moving forward!